\name{ALLs.chr22}
\docType{data}
\alias{ALLs.chr22}
\title{Chromosome 22 subset of ALL data for ALL1/AF4 and E2A/PBX1}
\description{
  This is a greatly reduced subset of the Chiaretti et al. ALL data
  set (available in its entirety as the Bioconductor ALL package). The
  data in this subset consist of microarrays from 15 different
  individuals with acute lymphoblastic leukemia (ALL). The data are
  further restricted to chromosome 22 only. This data set is intended
  for demonstration purposes only.
}
\usage{ALLs.chr22}
\format{An ExpressionSet with the following covariates:
  \itemize{
    \item{age}{The age of the patient in years.}
    \item{mol.biol}{The assigned molecular biology of the cancer (mainly
      for those with B-cell ALL). In this data set this is restricted to
      ALL1/AF4 and E2A/PBX1.}
  }
}
\source{The ALL Bioconductor data package}
\references{
  Sabina Chiaretti, Xiaochun Li, Robert Gentleman, Antonella Vitale,
  Marco Vignetti, Franco Mandelli, Jerome Ritz, and Robin Foa Gene
  expression profile of adult T-cell acute lymphocytic leukemia
  identifies distinct subsets of patients with different response to
  therapy and survival. Blood, 1 April 2004, Vol. 103, No. 7.
}
\keyword{datasets}
